Convert DashboadsPage to typescript
